rename: 폴더명 변경 및 구조 이동
