+++ src/page/registerPet/index/component/petHealth/disease/Step2.css.ts
