+++ src/page/registerPet/index/component/petHealth/disease/Step2Constant.ts
